requirements.txt
black==24.10.0
blinker==1.9.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
identify==2.6.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.28.0
Werkzeug==3.1.3
